chore: rename util.rs -> utils.rs
