+++ wonnx-cli/src/utils.rs
